test: delete system tests
The system tests being deleted are handwritten but they are for a client which is fully autogenerated. Issue googleapis/gapic-generator-python#1771 has been filed to evaluate whether these tests can be migrated to` gapic-showcase`, which is closest to where the actual development occurs. We need to delete these tests as it is blocking the migration to `google-cloud-python`, where we are trying to minimize handwritten code.